% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwhatsapp.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{List of emojis and corresponding descriptions.}
\format{A tibble with 3570 rows and 2 columns
- emoji: character representation of the emoji
- name: name of the emoji}
\source{
\url{https://github.com/hadley/emo/}
}
\usage{
emojis
}
\description{
A dataset containing emojis and corresponding descriptions. This is a subset
of the emojis provided by the emo package.
}
\keyword{datasets}
