% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji_lookup.R
\name{lookup_emoji}
\alias{lookup_emoji}
\alias{lookup_emoji.data.frame}
\alias{lookup_emoji.character}
\title{Lookup emojis from text}
\usage{
lookup_emoji(x, ...)

\method{lookup_emoji}{data.frame}(x, text_field = "text", ...)

\method{lookup_emoji}{character}(x, ...)
}
\arguments{
\item{x}{A character vector or data.frame.}

\item{...}{not currently used.}

\item{text_field}{the character name or numeric index of the source
`data.frame` indicating the variable to be read in as text.}
}
\description{
Takes a character string or data.frame with text, looks up all
  emoji characters in it and also returns their description. Supports the
  full unicode Emoji List v13.0 (see \link{emojis}).
}
