% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwhatsapp.R
\name{rwa_parse_time}
\alias{rwa_parse_time}
\title{Parse time}
\usage{
rwa_parse_time(time, format, tz)
}
\arguments{
\item{format}{Most formats are automatically detected. If you encounter
problems you can provide a custom format here. Refer to
\link[stringi]{stri_datetime_parse} for guidance.}

\item{tz}{A time zone for date conversion. Set NULL or "" for the default
time zone or a single string with a timezone identifier, see
\link[stringi]{stri_timezone_list}.}
}
\value{
a tibble
}
\description{
The history can be obtained going to the menu in a chat on the WhatsApp app,
choosing "more", then "Export chat".
}
\examples{
history <- system.file("extdata", "sample.txt", package = "rwhatsapp")
df <- rwa_read(history)
}
