% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwhatsapp.R
\name{rwa_read_lines}
\alias{rwa_read_lines}
\title{Read in files from supported formats}
\usage{
rwa_read_lines(x, verbose, ...)
}
\arguments{
\item{x}{Path to a txt or zip file of a WhatsApp history or the history
itself as character object.}

\item{verbose}{A logical flag indicating whether information should be
printed to the screen.}

\item{...}{Further arguments passed to \link[stringi]{stri_read_lines}.}
}
\value{
a tibble
}
\description{
The history can be obtained going to the menu in a chat on the WhatsApp app,
choosing "more", then "Export chat".
}
\examples{
history <- system.file("extdata", "sample.txt", package = "rwhatsapp")
df <- rwa_read(history)
}
