% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwhatsapp.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{List of emojis and corresponding descriptions.}
\format{A tibble with 3570 rows and 3 columns: \itemize{
  \item emoji character representation of the emoji
  \item name of the emoji
  \item hex_runes hexadecimal representations of emoji
}}
\source{
\url{https://github.com/hadley/emo/}
}
\usage{
emojis
}
\description{
A dataset containing emojis and corresponding descriptions. This is a subset
of the emojis provided by the emo package.
}
\details{
\code{hex_runes} can be used to easily look up image files of
  emojis.
}
\keyword{datasets}
